# e-commerce_backend/openapi-spec.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/auth/register": {
            "post": {
                "operationId": "AuthController_register",
                "summary": "Register a new user",
                "description": "Create a new user account with email verification",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "400": {
                        "description": "Invalid input or user already exists"
                    }
                },
                "tags": [
                    "Authentication"
                ]
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "operationId": "AuthController_login",
                "summary": "Login user",
                "description": "Authenticate user and receive JWT tokens in cookies",
                "parameters": [
                    {
                        "name": "user-agent",
                        "required": true,
                        "in": "header",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful, tokens set in cookies"
                    },
                    "401": {
                        "description": "Invalid credentials or seller pending verification"
                    }
                },
                "tags": [
                    "Authentication"
                ]
            }
        },
...
        "/api/v1/auth/refresh": {
            "post": {
                "operationId": "AuthController_refreshToken",
                "parameters": [
                    {
                        "name": "user-agent",
                        "required": true,
                        "in": "header",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Authentication"
                ]
            }
        },
        "/api/v1/auth/forgot-password": {
            "post": {
                "operationId": "AuthController_forgotPassword",
                "parameters": [
                    {
                        "name": "user-agent",
                        "required": true,
                        "in": "header",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ForgotPasswordDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Authentication"
                ]
            }
        },
        "/api/v1/auth/verify-otp": {
            "post": {
                "operationId": "AuthController_verifyOtp",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VerifyOtpDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Authentication"
                ]
            }
        },
        "/api/v1/auth/reset-password": {
            "post": {
                "operationId": "AuthController_resetPassword",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResetPasswordDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Authentication"
                ]
            }
        },
...
        "/api/v1/auth/google/callback": {
            "get": {
                "operationId": "AuthController_googleAuthCallback",
                "parameters": [
                    {
                        "name": "user-agent",
                        "required": true,
                        "in": "header",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Authentication"
                ]
            }
        },
...
    },
    "info": {
        "title": "E-Commerce Platform API",
        "description": "Complete API documentation for the E-Commerce platform with authentication, products, orders, notifications, and seller management.",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [
        {
            "url": "http://localhost:4002",
            "description": "Development Server"
        },
        {
            "url": "http://localhost:4002/api/v1",
            "description": "API v1"
        }
    ],
    "components": {
        "securitySchemes": {
            "JWT-auth": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http",
                "description": "Enter JWT token"
            },
            "cookie": {
                "type": "apiKey",
                "in": "cookie",
                "name": "access_token",
                "description": "JWT token stored in HTTP-only cookie"
            }
        },
        "schemas": {
...
            "RegisterDto": {
                "type": "object",
                "properties": {}
            },
            "LoginDto": {
                "type": "object",
                "properties": {}
            },
            "ForgotPasswordDto": {
                "type": "object",
                "properties": {}
            },
            "VerifyOtpDto": {
                "type": "object",
                "properties": {}
            },
            "ResetPasswordDto": {
                "type": "object",
                "properties": {}
            },
...
        }
    }
}